% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residplot.R
\name{residplot}
\alias{residplot}
\title{residplot}
\usage{
residplot(fit, nbreaks = 10)
}
\arguments{
\item{fit}{the result of regression}

\item{nbreaks}{the number of group}
}
\value{
a random sample of size \code{n}
}
\description{
an example function  generating Student residuals histogram
}
\examples{
states<-as.data.frame(state.x77[,c("Murder","Population","Illiteracy","Income","Frost")])
fit<-lm(Murder~ Population + Illiteracy + Income + Frost,data=states)
residplot(fit)
}
