% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{shrinkage}
\alias{shrinkage}
\title{shrinkage}
\usage{
shrinkage(fit, k = 10)
}
\arguments{
\item{fit}{the result of regression}

\item{k}{the number of k}
}
\value{
a random sample of size \code{n}
}
\description{
an example function achieving k fold Cross-Validation for R square value.
}
\examples{
states<-as.data.frame(state.x77[,c("Murder","Population","Illiteracy","Income","Frost")])
fit<-lm(Murder~ Population + Illiteracy + Income + Frost,data=states)
shrinkage(fit)
}
